声鉴-自定义昵称

+++ app/controllers/m/UsersController.php
@@ -135,6 +135,7 @@ class UsersController extends BaseController
     {
         $user = $this->currentUser();
         $sex = $this->params('sex');
+        $nickname = $this->params('nickname');
         $all_read_test = ['可能我只是你生命里的一个过客，但你不会遇见第二个我', '当我们搬开别人脚下的绊脚石时,也许恰恰是在为自己铺路',
             '如果你看到面前的阴影，别怕，那是因为你的背后有阳光！', '玫瑰你的，巧克力你的，钻石你的。你，是我的！',
             '为什么我的眼里常含泪水，因为我的同桌老给我丢人。', '最怕那些对我很重要的人连句再见都没有就突然消失不见',
@@ -143,6 +144,7 @@ class UsersController extends BaseController
 
         $read_text_index = array_rand($all_read_test);
         $this->view->read_text = $all_read_test[$read_text_index];
+        $this->view->nickname = $nickname;
         $this->view->sex = $sex;
         $this->view->user = json_encode($user->toChatJson(), JSON_UNESCAPED_UNICODE);
     }
@@ -151,7 +153,9 @@ class UsersController extends BaseController
     {
         $user = $this->currentUser();
         $sex = $this->params('sex');
+        $nickname = $this->params('nickname');
         $this->view->sex = $sex;
+        $this->view->nickname = $nickname;
         $this->view->user = json_encode($user->toChatJson(), JSON_UNESCAPED_UNICODE);
     }
 
@@ -285,12 +289,16 @@ class UsersController extends BaseController
                     $img_path = str_replace('../../..', '', $new_file);
                     $img_files[] = $img_path;
                     $res = \Albums::uploadImage($user, $img_files);
-//                    $image_result = $user->saveImageToPhone();
+                    $image_result = $this->saveImageToPhone();
                     if ($res) {
-                        if (file_exists($source_filename)) {
-                            unlink($source_filename);
+                        if ($image_result) {
+                            if (file_exists($source_filename)) {
+                                unlink($source_filename);
+                            }
+                            return $this->renderJSON(ERROR_CODE_SUCCESS, '图片已成功保存到个人资料相册');
+                        } else {
+                            return $this->renderJSON(ERROR_CODE_FAIL, '图片保存失败');
                         }
-                        return $this->renderJSON(ERROR_CODE_SUCCESS, '图片已成功保存到个人资料相册');
                     } else {
                         return $this->renderJSON(ERROR_CODE_FAIL, '图片上传失败');
                     }
@@ -306,5 +314,47 @@ class UsersController extends BaseController
             //文件错误
             return $this->renderJSON(ERROR_CODE_FAIL, '文件错误');
         }
+    }
+
+    function saveImageToPhone($file = '', $timeout = 60)
+    {
+        $album = \Albums::findFirst([
+            'conditions' => 'user_id=:user_id: and auth_status=:auth_status:',
+            'bind' => ['user_id' => $this->currentUserId(), 'auth_status' => AUTH_WAIT],
+            'order' => 'created_at desc'
+        ]);
+        $url = $album->image_url;
+        $file = empty($file) ? pathinfo($url, PATHINFO_BASENAME) : $file;
+        $dir = pathinfo($file, PATHINFO_DIRNAME);
+        !is_dir($dir) && @mkdir($dir, 0755, true);
+        $url = str_replace(" ", "%20", $url);
+
+        if (function_exists('curl_init')) {
+            $ch = curl_init();
+            curl_setopt($ch, CURLOPT_URL, $url);
+            curl_setopt($ch, CURLOPT_TIMEOUT, $timeout);
+            curl_setopt($ch, CURLOPT_RETURNTRANSFER, TRUE);
+            $temp = curl_exec($ch);
+            if (@file_put_contents($file, $temp) && !curl_error($ch)) {
+                return true;
+            } else {
+                return false;
+            }
+        } else {
+            $opts = array(
+                "http" => array(
+                    "method" => "GET",
+                    "header" => "",
+                    "timeout" => $timeout)
+            );
+            $context = stream_context_create($opts);
+            if (@copy($url, $file, $context)) {
+                //$http_response_header
+                return true;
+            } else {
+                return false;
+            }
+        }
+
     }
 }
